PR295レビュー指摘を反映してclaims取得を共通化
