+++ docs/admin_spec.md
@@ -148,13 +148,13 @@ OpenAPI スコープで表現できない条件付き認可が必要な場合の
 
 | user_id | 名前 | 学籍番号 | is_member | 役職 claim |
 |---------|------|----------|-----------|-----------|
-| `11111111-1111-1111-1111-111111111111` | 田中太郎 | 20230001 | `true` | `account` |
-| `22222222-2222-2222-2222-222222222222` | 佐藤花子 | 20230002 | `true` | なし |
-| `33333333-3333-3333-3333-333333333333` | 山田次郎 | 20230003 | `true` | `infra` |
-| `44444444-4444-4444-4444-444444444444` | 鈴木美咲 | 20230004 | `false` | なし |
-| `55555555-5555-5555-5555-555555555555` | 高橋健太 | 20230005 | `true` | なし |
-| `66666666-6666-6666-6666-666666666666` | 伊藤一真 | 20230006 | `false` | なし |
-| `77777777-7777-7777-7777-777777777777` | 中村直樹 | 20230007 | `false` | `infra` |
+| `11111111-1111-1111-1111-111111111111` | 田中太郎 | aa230001 | `true` | `account` |
+| `22222222-2222-2222-2222-222222222222` | 佐藤花子 | bp230002 | `true` | なし |
+| `33333333-3333-3333-3333-333333333333` | 山田次郎 | cy230003 | `true` | `infra` |
+| `44444444-4444-4444-4444-444444444444` | 鈴木美咲 | z2300004 | `false` | なし |
+| `55555555-5555-5555-5555-555555555555` | 高橋健太 | ab230005 | `true` | なし |
+| `66666666-6666-6666-6666-666666666666` | 伊藤一真 | bq230006 | `false` | なし |
+| `77777777-7777-7777-7777-777777777777` | 中村直樹 | dp230007 | `false` | `infra` |
 
 ---
