+++ schema/test/insert_test.sql
@@ -18,7 +18,7 @@ INSERT INTO user_profiles (id, user_id, username, school_grade, icon_url, discor
   (UUID_TO_BIN('dddd4444-4444-4444-4444-444444444444'), UUID_TO_BIN('44444444-4444-4444-4444-444444444444'), '鈴木美咲', 1, 'https://example.com/icons/suzuki.jpg', 'suzuki012', '2026-12-31', false, 'AI・機械学習研究中', '人工知能と機械学習の研究をしています。Python、TensorFlow、PyTorchを使ってモデル開発をしています。'),
   (UUID_TO_BIN('eeee5555-5555-5555-5555-555555555555'), UUID_TO_BIN('55555555-5555-5555-5555-555555555555'), '高橋健太', 3, 'https://example.com/icons/takahashi.jpg', 'takahashi345', '2024-12-31', true, 'フロントエンド開発者', 'モダンなWebフロントエンド開発が専門です。Vue.js、TypeScript、アニメーションライブラリが得意です。'),
   (UUID_TO_BIN('ffff6666-6666-6666-6666-666666666666'), UUID_TO_BIN('66666666-6666-6666-6666-666666666666'), '伊藤一真', 2, 'https://example.com/icons/ito.jpg', 'ito678', '2025-12-31', false, '休部中メンバー', '現在は休部中で、個人でWeb制作を進めています。'),
-  (UUID_TO_BIN('ffff7777-7777-7777-7777-777777777777'), UUID_TO_BIN('77777777-7777-7777-7777-777777777777'), '中村直樹', 4, 'https://example.com/icons/nakamura.jpg', 'nakamura901', '2025-03-31', false, '運営サポート', '運営タスクを中心にサポートしています。');
+  (UUID_TO_BIN('77777777-aaaa-7777-7777-777777777777'), UUID_TO_BIN('77777777-7777-7777-7777-777777777777'), '中村直樹', 4, 'https://example.com/icons/nakamura.jpg', 'nakamura901', '2025-03-31', false, '運営サポート', '運営タスクを中心にサポートしています。');
 
 -- イベントデータの投入
 INSERT INTO events (id, name, calendar_view, description) VALUES
